refactor(api): separate mappers from services
