+++ services/api/src/main/java/dev/dictum/api/content/mapper/PostApiMapper.java
@@ -1,4 +1,4 @@
-package dev.dictum.api.content.service;
+package dev.dictum.api.content.mapper;
 
 import dev.dictum.api.content.model.vo.PostState;
 import dev.dictum.api.generated.model.PostResponse;
@@ -8,7 +8,7 @@ import org.mapstruct.Mapper;
 import org.mapstruct.MappingConstants;
 
 @Mapper(componentModel = MappingConstants.ComponentModel.SPRING)
-interface PostApiMapper {
+public interface PostApiMapper {
 
   PostSummary toSummary(PostState state);
