+++ services/api/src/main/java/dev/dictum/api/site/mapper/SiteSettingsApiMapper.java
@@ -1,4 +1,4 @@
-package dev.dictum.api.site.service;
+package dev.dictum.api.site.mapper;
 
 import dev.dictum.api.generated.model.SiteSettingsResponse;
 import dev.dictum.api.site.model.vo.SiteSettingsState;
@@ -6,7 +6,7 @@ import org.mapstruct.Mapper;
 import org.mapstruct.MappingConstants;
 
 @Mapper(componentModel = MappingConstants.ComponentModel.SPRING)
-interface SiteSettingsApiMapper {
+public interface SiteSettingsApiMapper {
 
   SiteSettingsResponse toResponse(SiteSettingsState state);
 }
