+++ services/api/src/main/java/dev/dictum/api/site/service/SiteSettingsQueryService.java
@@ -1,6 +1,7 @@
 package dev.dictum.api.site.service;
 
 import dev.dictum.api.generated.model.SiteSettingsResponse;
+import dev.dictum.api.site.mapper.SiteSettingsApiMapper;
 import dev.dictum.api.site.repository.InMemorySiteSettingsStore;
 import org.springframework.stereotype.Service;
